feat: Add New Method 'UpdateStashBalanceAmount' and 'Internal Methods'

+++ Core/Services/CurrencyAccountService.cs
@@ -5,11 +5,13 @@ using Core.Domain.RepositoryContracts;
 using Core.DTO;
 using Core.DTO.CurrencyAccountDTO;
 using Core.DTO.TransactionDTO;
+using Core.Helpers;
 using Core.ServiceContracts;
 using Microsoft.AspNetCore.Identity;
 
 namespace Core.Services;
 
+// todo: check response of 'SaveChangesAsync'
 public class CurrencyAccountService : ICurrencyAccountService
 {
     private readonly ICurrencyAccountRepository _accountRepository;
@@ -25,9 +27,8 @@ public class CurrencyAccountService : ICurrencyAccountService
         _transactionService = transactionService;
     }
 
-    public async Task<(bool isValid, string? message, CurrencyAccountResponse? obj)> AddCurrencyAccount(CurrencyAccountAddRequest? currencyAccountAddRequest, ClaimsPrincipal userClaims)
+    public async Task<(bool isValid, string? message, CurrencyAccountAddResponse? obj)> AddCurrencyAccount(CurrencyAccountAddRequest? currencyAccountAddRequest, ClaimsPrincipal userClaims)
     {
-        // 'CurrencyAccountAddRequest' is Null //
         ArgumentNullException.ThrowIfNull(currencyAccountAddRequest,"The 'CurrencyAccountAddRequest' object parameter is Null");
         
         var currency = await _currencyService.GetCurrencyByCurrencyType(currencyAccountAddRequest.CurrencyType);
@@ -40,14 +41,18 @@ public class CurrencyAccountService : ICurrencyAccountService
             AccountNumber = currencyAccount.Number,
             Money = new MoneyRequest(){Amount = currencyAccountAddRequest.MoneyToOpenAccount.Amount!.Value, CurrencyType = currencyAccountAddRequest.MoneyToOpenAccount.CurrencyType}
         };
-        var (isValid, message, _) = await _transactionService.Value.AddDepositTransaction(transactionAddRequest, userClaims);
+        
+        var (isValid, message, tranasctionResponse) = await _transactionService.Value.AddOpenAccountDepositTransaction(transactionAddRequest, currencyAccountAddRequest);
         if (!isValid)
             return (false, message, null);
         
         var currencyAccountReturned = await _accountRepository.AddCurrencyAccountAsync(currencyAccount);
-        await _accountRepository.SaveChangesAsync();
+        
+        await UpdateStashBalanceAmount(currencyAccount, tranasctionResponse!.FromAccountChangeAmount, (val1, val2) => val1 + val2);
+        var numberOfRowsAffected = await _accountRepository.SaveChangesAsync();
+        if (!(numberOfRowsAffected >= 2)) return (false, "The Request Has Not Been Done Completely, Try Again", null);
 
-        var currencyAccountResponse = currencyAccountReturned.ToCurrencyAccountResponse();
+        var currencyAccountResponse = currencyAccountReturned.ToCurrencyAccountResponse(tranasctionResponse.Id);
         return (true, null, currencyAccountResponse);
     }   
     
@@ -56,35 +61,31 @@ public class CurrencyAccountService : ICurrencyAccountService
         var user = await _userManager.GetUserAsync(userClaims);
         
         List<CurrencyAccount> currencyAccounts;
-        if (userClaims.IsInRole("Admin"))
-            return GetAllCurrencyAccountsInternal();
-       
-        currencyAccounts = _accountRepository.GetCurrencyAccounts()
-                                             .Where(property => property.OwnerID == user.Id)
-                                             .ToList();
+        if (userClaims.IsInRole(Constants.AdminRole))
+            return await GetAllCurrencyAccountsInternal();
+
+        currencyAccounts = await _accountRepository.GetAllCurrencyAccountsByUserAsync(user!.Id);
         
         var currencyAccountResponses = currencyAccounts.Select(accountItem => accountItem.ToCurrencyAccountResponse()).ToList();
         return currencyAccountResponses;
     }
     
-    public List<CurrencyAccountResponse> GetAllCurrencyAccountsInternal()
+    public async Task<List<CurrencyAccountResponse>> GetAllCurrencyAccountsInternal()
     {
-        var currencyAccounts = _accountRepository.GetCurrencyAccounts().ToList();
+        var currencyAccounts = await _accountRepository.GetAllCurrencyAccountsAsync();
         var currencyAccountResponses = currencyAccounts.Select(accountItem => accountItem.ToCurrencyAccountResponse()).ToList();
         return currencyAccountResponses;
     }
     
     public async Task<(bool isValid, string? message, CurrencyAccountResponse? obj)> GetCurrencyAccountByNumber(string? number, ClaimsPrincipal userClaims)
     {
-        // if 'number' is null
         ArgumentNullException.ThrowIfNull(number,"The CurrencyAccount'number' parameter is Null");
         
-        if (userClaims.IsInRole("Admin"))
-            return GetCurrencyAccountByNumberInternal(number);
+        if (userClaims.IsInRole(Constants.AdminRole))
+            return await GetCurrencyAccountByNumberInternal(number);
         
         var user = await _userManager.GetUserAsync(userClaims);
-        var currencyAccount = _accountRepository.GetCurrencyAccounts()
-                                                .FirstOrDefault(property => property.Number == number);
+        var currencyAccount = await _accountRepository.GetCurrencyAccountByNumberAsync(number);
 
         // if 'number' doesn't exist in 'currencyAccounts list' 
         if (currencyAccount == null) 
@@ -97,51 +98,96 @@ public class CurrencyAccountService : ICurrencyAccountService
         return (true, null, currencyAccountResponse);
     }
     
-    public (bool isValid, string? message, CurrencyAccountResponse? obj) GetCurrencyAccountByNumberInternal(string? number)
+    public async Task<(bool isValid, string? message, CurrencyAccountResponse? obj)> GetCurrencyAccountByNumberInternal(string? number)
     {
-        // if 'number' is null
         ArgumentNullException.ThrowIfNull(number,"The CurrencyAccount'number' parameter is Null");
 
-        var currencyAccount = _accountRepository.GetCurrencyAccounts()
-                                                .FirstOrDefault(property => property.Number == number);
+        var currencyAccount = await _accountRepository.GetCurrencyAccountByNumberAsync(number);
         
         // if 'number' doesn't exist in 'currencyAccounts list' 
         if (currencyAccount == null) 
             return (false, null, null);
-        
+
         var currencyAccountResponse = currencyAccount.ToCurrencyAccountResponse();
         return (true, null, currencyAccountResponse);
+    }   
+    
+    public async Task<(bool isValid, string? message, CurrencyAccount? obj)> GetCurrencyAccountByNumberWithNavigationInternal(string? number)
+    {
+        ArgumentNullException.ThrowIfNull(number,"The CurrencyAccount'number' parameter is Null");
+
+        var currencyAccount = await _accountRepository.GetCurrencyAccountByNumberAsync(number);
+        
+        // if 'number' doesn't exist in 'currencyAccounts list' 
+        if (currencyAccount == null) 
+            return (false, null, null);
+
+        return (true, null, currencyAccount);
     }
     
-    public async Task<(bool isValid, bool isFound, string? message)> DeleteCurrencyAccount(string? number, ClaimsPrincipal userClaims)
+    public async Task<(bool isValid, bool isFound, string? message)> DeleteCurrencyAccountByNumber(string? number, ClaimsPrincipal userClaims)
     {
-        // if 'number' is null
         ArgumentNullException.ThrowIfNull(number,"The CurrencyAccount'Number' parameter is Null");
-
-        var currencyAccount =  _accountRepository.GetCurrencyAccounts()
-                                                 .FirstOrDefault(property => property.Number == number);
+        
+        if (userClaims.IsInRole(Constants.AdminRole))
+            return await DeleteCurrencyAccountByNumberInternal(number);
+        
+        var user = await _userManager.GetUserAsync(userClaims);
+        var currencyAccount = await _accountRepository.GetCurrencyAccountByNumberAsync(number);
         
         // if 'Number' is invalid (doesn't exist)
         if (currencyAccount == null)
             return (false, false, null);
     
-        bool result = _accountRepository.DeleteCurrencyAccount(currencyAccount);
-        await _accountRepository.SaveChangesAsync();
+        if (currencyAccount.OwnerID != user!.Id)
+            return (false, true, "This Account Number Doesn't Belong To You");
+        
+        _accountRepository.DeleteCurrencyAccount(currencyAccount);
+        var numberOfRowsAffected = await _accountRepository.SaveChangesAsync();
+        if (!(numberOfRowsAffected > 0)) return (false, true, "The Request Has Not Been Done Completely, Try Again");
 
-        return (result, true, "The Deletion Has Not Been Successful");
+        return (true, true, "The Deletion Has Not Been Successful");
+    }
+
+    public async Task<(bool isValid, bool isFound, string? message)> DeleteCurrencyAccountByNumberInternal(string? number)
+    {
+        ArgumentNullException.ThrowIfNull(number,"The 'number' parameter is Null");
+
+        var currencyAccount = await _accountRepository.GetCurrencyAccountByNumberAsync(number);
+        
+        // if 'number' doesn't exist in 'currencyAccounts list' 
+        if (currencyAccount == null) 
+            return (false, false, null);
+        
+        _accountRepository.DeleteCurrencyAccount(currencyAccount);
+        var numberOfRowsAffected = await _accountRepository.SaveChangesAsync();
+        if (!(numberOfRowsAffected > 0)) return (false, true, "The Request Has Not Been Done Completely, Try Again");
+
+        return (true, true, null);
     }
     
-    
-    public async Task<CurrencyAccountResponse?> UpdateBalanceAmount(CurrencyAccount? currencyAccount, decimal? amount, Func<decimal,decimal,decimal> calculationFunc)
+    public async Task<CurrencyAccountResponse?> UpdateStashBalanceAmount(CurrencyAccount? currencyAccount, decimal? amount, Func<decimal,decimal,decimal> calculationFunc, bool isTransactionConfirmed = false)
     {
-        // if 'currencyAccount' is null
         ArgumentNullException.ThrowIfNull(currencyAccount, "The 'currencyAccount' parameter is Null");
-        
-        // if 'amount' is null
         ArgumentNullException.ThrowIfNull(amount, "The 'amount' parameter is Null");
         
-        var updatedCurrencyAccount = _accountRepository.UpdateBalanceAmount(currencyAccount, amount.Value, calculationFunc);
-        await _accountRepository.SaveChangesAsync();
+        var updatedCurrencyAccount = _accountRepository.UpdateStashBalanceAmount(currencyAccount, amount.Value, calculationFunc);
+        
+        // var numberOfRowsAffected = await _accountRepository.SaveChangesAsync();
+        // if (!(numberOfRowsAffected > 0)) return null;
+        
+        return updatedCurrencyAccount.ToCurrencyAccountResponse();
+    }    
+    
+    public async Task<CurrencyAccountResponse?> UpdateBalanceAmount(CurrencyAccount? currencyAccount, decimal? amount, Func<decimal,decimal,decimal> calculationFunc, bool isTransactionConfirmed = false)
+    {
+        ArgumentNullException.ThrowIfNull(currencyAccount, "The 'currencyAccount' parameter is Null");
+        ArgumentNullException.ThrowIfNull(amount, "The 'amount' parameter is Null");
+
+        var updatedCurrencyAccount =_accountRepository.UpdateBalanceAmount(currencyAccount, amount.Value, calculationFunc);
+        
+        // var numberOfRowsAffected = await _accountRepository.SaveChangesAsync();
+        // if (!(numberOfRowsAffected > 0)) return null;
         
         return updatedCurrencyAccount.ToCurrencyAccountResponse();
     }
